feat: init customer secrets
