+++ packages/spacecat-shared-google-client/package.json
@@ -34,6 +34,7 @@
     "@adobe/helix-universal": "4.5.2",
     "@adobe/spacecat-shared-http-utils": "1.1.4",
     "@adobe/spacecat-shared-utils": "1.14.5",
+    "@aws-sdk/client-secrets-manager": "3.556.0",
     "google-auth-library": "9.8.0",
     "googleapis": "134.0.0"
   },
